feat: Implementação logica para tipo de transação e conversão da taxa de cambio

+++ transacao-consumer-api/src/main/java/br/com/ntt/transacao/consumer/domain/entities/transacao/Transacao.java
@@ -79,7 +79,8 @@ public class Transacao {
         this.status = statusTransacao;
     }
 
-    public void atualizaTaxaDeCambio(BigDecimal cotacaoVenda) {
-        this.taxaCambio = cotacaoVenda;
+    public void atualizaTaxaDeCambio(Double cotacaoVenda) {
+        this.taxaCambio = BigDecimal.valueOf(cotacaoVenda);
+
     }
 }
